+++ tests/epyccel/test_epyccel_openmp.py
@@ -353,7 +353,7 @@ def test_omp_nowait(language):
     f1 = epyccel(openmp.omp_nowait, fflags = '-Wall', accelerators=['openmp'], language=language)
     set_num_threads = epyccel(openmp.set_num_threads, fflags = '-Wall', accelerators=['openmp'], language=language)
     set_num_threads(4)
-    x = random.randint(20, size=(1000))
+    x = random.randint(20, size=(1000), dtype=int)
     y = np.zeros((1000,), dtype=int)
     z = np.zeros((1000,))
     f1(x, y, z)
@@ -366,7 +366,7 @@ def test_omp_arraysum(language):
     f1 = epyccel(openmp.omp_arraysum, fflags = '-Wall', accelerators=['openmp'], language=language)
     set_num_threads = epyccel(openmp.set_num_threads, fflags = '-Wall', accelerators=['openmp'], language=language)
     set_num_threads(4)
-    x = random.randint(20, size=(5))
+    x = random.randint(20, size=(5), dtype=int)
 
     assert f1(x) == np.sum(x)
 
@@ -375,7 +375,7 @@ def test_omp_arraysum_combined(language):
     f1 = epyccel(openmp.omp_arraysum_combined, fflags = '-Wall', accelerators=['openmp'], language=language)
     set_num_threads = epyccel(openmp.set_num_threads, fflags = '-Wall', accelerators=['openmp'], language=language)
     set_num_threads(4)
-    x = random.randint(20, size=(5))
+    x = random.randint(20, size=(5), dtype=int)
 
     assert f1(x) == np.sum(x)
 
@@ -393,7 +393,7 @@ def test_omp_arraysum_single(language):
     f1 = epyccel(openmp.omp_arraysum_single, fflags = '-Wall', accelerators=['openmp'], language=language)
     set_num_threads = epyccel(openmp.set_num_threads, fflags = '-Wall', accelerators=['openmp'], language=language)
     set_num_threads(2)
-    x = random.randint(20, size=(10))
+    x = random.randint(20, size=(10), dtype=int)
 
     assert f1(x) == np.sum(x)
 
@@ -446,11 +446,11 @@ def test_omp_long_line(language):
     f1 = epyccel(openmp.omp_long_line, fflags = '-Wall', accelerators=['openmp'], language=language)
     set_num_threads = epyccel(openmp.set_num_threads, fflags = '-Wall', accelerators=['openmp'], language=language)
     set_num_threads(4)
-    x1 = random.randint(20, size=(5))
-    x2 = random.randint(20, size=(5))
-    x3 = random.randint(20, size=(5))
-    x4 = random.randint(20, size=(5))
-    x5 = random.randint(20, size=(5))
+    x1 = random.randint(20, size=(5), dtype=int)
+    x2 = random.randint(20, size=(5), dtype=int)
+    x3 = random.randint(20, size=(5), dtype=int)
+    x4 = random.randint(20, size=(5), dtype=int)
+    x5 = random.randint(20, size=(5), dtype=int)
 
     assert f1(x1,x2,x3,x4,x5) == np.sum(x1+x2+x3+x4+x5)
